hunger-games-backend/src/main/java/ru/itmo/hungergames/service/impl/ResourceServiceImpl.java
package ru.itmo.hungergames.service.impl;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import ru.itmo.hungergames.model.entity.order.Resource;
import ru.itmo.hungergames.repository.ResourceRepository;
import ru.itmo.hungergames.service.ResourceService;

import java.util.List;

@Service
@Transactional(readOnly = true)
public class ResourceServiceImpl implements ResourceService {
    private final ResourceRepository resourceRepository;

    @Autowired
    public ResourceServiceImpl(ResourceRepository resourceRepository)  {
        this.resourceRepository = resourceRepository;
    }

    @Override
    public List<Resource> getAllResources() {
        return resourceRepository.findAll();
    }
}
